Finished documenting the project.
Finals testing should be made before submitting it.

(Friday, 18/08/23)

+++ FileHandling/readFromFile.h
@@ -1,6 +1,6 @@
 /*
  * @author Ofir Paz
- * @Version (27/04/2023)
+ * @version (18/08/2023)
  *
  * Header file for the readFromFile.c file.
  */
@@ -22,7 +22,7 @@
  * @return  The length of the line if it has been read successfully, EOF (-1) if the line
  *          was not read successfully, -2 if the file was not opened.
  */
-ssize_t readNextLineFromFile(const char *file_name, const char *fileType, char **line);
+int readNextLineFromFile(const char *file_name, const char *fileType, char **line);
 
 /* ---------------------------------------- */
